modified global recs to account for local vs remote

+++ globalreq.js
@@ -1,20 +1,20 @@
-var redis = require("redis");
+var redis = require("redis"),
+    config = require('./config.js').config();
+
 
 module.exports = function() {
-  client = redis.createClient();
-  // client = redis.createClient(6379,'127.0.0.1');
-  // client = redis.createClient(12000, process.env.REDIS_HOST);
-  // client.auth(process.env.REDIS_AUTH, function (err) {
-  //  if (err) { throw err; }
-  //  // You are now connected to your redis.
-  // });
-// client = redis.createClient(6379,'nodejitsudb6625264643.redis.irstack.com');
-  // client.auth("nodejitsudb6625264643.redis.irstack.com:f327cfe980c971946e80b8e975fbebb4", function(err, reply) {
-  //   if (err) {  console.log("Error: "+err); }
-  //   else { console.log("should be connected. Reply: "+reply); }
-  // });
-  client.flushdb();
-  // client.on('ready', function(){
-
-  // });
+  if (config.localSetup ===  true){
+    client = redis.createClient(config.localRedisPort,config.localRedisURL);
+  } else {
+    client = redis.createClient(config.remoteRedisPort, config.remoteRedisURL);
+    client.auth(config.remoteRedisAuth, function (err) {
+     if (err) { throw err; }
+     // You are now connected to your redis.
+    });
+  }
+  if (config.flushDBsOnStart){
+    client.flushdb();
+  }
+  client.on('ready', function(){
+  });
 };
